few-shot/requirements.txt
torch>=2.0.0 --index-url https://download.pytorch.org/whl/cu121
transformers>=4.30.0
pyyaml>=6.0
accelerate>=0.26.0
bitsandbytes>=0.41.0
protobuf>=4.25.2
sentencepiece>=0.2.0